refactor(engine/types): move TransactionId to root _types.py
