+++ src/prisma/engine/__init__.py
@@ -1,5 +1,5 @@
-from ._types import TransactionId as TransactionId
 from .errors import *
+from .._types import TransactionId as TransactionId
 from ._abstract import BaseAbstractEngine as BaseAbstractEngine
 
 try:
