--- src/prisma/engine/_types.py
@@ -1,3 +1,0 @@
-from typing_extensions import NewType
-
-TransactionId = NewType('TransactionId', str)
